did alot of stuff idk
