+++ backend/server.js
@@ -2,6 +2,7 @@ const path = require('path');
 const express = require('express');
 const apiRouter = require('./apiRouter.js');
 
+
 const app = express();
 const PORT = 3000;
