+++ frontend/App.jsx
@@ -1,128 +1,29 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import handleSubmit from '../backend/request';
 import { useState } from 'react';
+import Login from './src/Login.jsx';
+import Calculator from './Calculator.jsx';
 
 const App = () => {
-  const [login, setLogin] = useState(true);
+  const [login, setLogin] = useState(false);
 
   const loginfunc = () => {
     setLogin((pastState) => !pastState);
-    console.log(login);
-  };
-
-  const newUserEntry =  async () => {
-    // users input username
-    const username = document.querySelector('#username').value;
-
-    // user input password
-    const password = document.querySelector('#password').value;
-    console.log(username, password);
-
-    try {
-      const myHeaders = new Headers();
-      myHeaders.append('Content-Type', 'application/json');
-
-      const promise = await fetch('/api/saveuser', {
-        method: 'POST',
-        body: JSON.stringify({ username, password }),
-        headers: myHeaders
-      });
-      const data = await response.json();
-      console.log('data returned: ', data);
-    } catch (error) {
-      console.log('Error in handleSubmit function in the App.jsx', error);
-    }
-
-
-
+    //console.log(login);
   };
 
   if (!login) {
     return (
-      <main>
-        <h1>Yeti Crab Mortgage Calculator</h1>
-        <div>
-          <input
-            id='loanAmmount'
-            type='number'
-            name='loanAmt'
-            placeholder='Loan Amount'
-          ></input>
-        </div>
-        <br></br>
-        <div>
-          <select name='loanTerm' id='loanTerm'>
-            <option value=''>Select Loan Term</option>
-            <option value='15'>15 Years</option>
-            <option value='20'>20 Years</option>
-            <option value='30'>30 Years</option>
-          </select>
-        </div>
-        <br></br>
-        <div>
-          <input
-            type='number'
-            name='interestRate'
-            placeholder='Interest Rate'
-            id='interestRate'
-          ></input>
-        </div>
-        <br></br>
-        <div>
-          <input
-            type='number'
-            name='homeInsurance'
-            placeholder='Home Insurance'
-            id='homeInsurance'
-          ></input>
-        </div>
-        <br></br>
-        <button type='submit' className='button' onClick={handleSubmit}>
-          {' '}
-          Submit{' '}
-        </button>
+      <div>
         <button onClick={loginfunc}>Go To Login </button>
-      </main>
+        <Calculator />
+      </div>
     );
   } else {
     return (
-      <div className='pageContianer'>
-        <div id='titleContainer'>
-          <div id='icon'></div>
-          <h3>Yeti Crab Mortgage Calculator</h3>
-        </div>
-
-        <div className='loginContainer'>
-          <h2>Create An Account</h2>
-
-          <div className='userlogin'>
-            <label for='username'>Enter Your New Username</label>
-            <br></br>
-            <input
-              id='username'
-              name='username'
-              placeholder="Ex: 'Jayson123'/'Goku' "
-            ></input>
-          </div>
-          <div className='passlogin'>
-            <label for='password'>Enter Your New Password</label>
-            <br></br>
-            <input
-              id='password'
-              name='password'
-              placeholder='Numbers & Characters'
-            ></input>
-            <br></br>
-            <button className='userSubmit' onClick={newUserEntry}>
-              {' '}
-              Become a new User
-            </button>
-          </div>
-
-          <button onClick={loginfunc}>Back to Homepage </button>
-        </div>
-        <div className='mortgageImage'></div>
+      <div>
+        <button onClick={loginfunc}>Back to Homepage </button>
+        <Login />
       </div>
     );
   }
