+++ package.json
@@ -41,6 +41,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.5.0",
+    "bcrypt": "^5.1.1",
     "express": "^4.21.2",
     "pg": "^8.13.1",
     "postgres": "^3.4.5",
